+++ packages/docs/src/stories/Box.stories.tsx
@@ -1,15 +1,11 @@
-import type { StoryObj, Meta } from '@storybook/react'
-import { Box, BoxProps, Text } from '@guaribas-ui/react'
+import type { Meta, StoryObj } from '@storybook/react'
+import { Box, BoxProps, Text } from '@healthcare-ui/react'
 
 export default {
   title: 'Surfaces/Box',
   component: Box,
   args: {
-    children: (
-      <>
-        <Text>Testando o elemento em tela </Text>
-      </>
-    ),
+    children: <Text>Testando o elemento Box</Text>,
   },
   argTypes: {
     children: {