+++ packages/docs/src/stories/TextInput.stories.tsx
@@ -1,12 +1,10 @@
-import type { StoryObj, Meta } from '@storybook/react'
-import { Box, TextInput, TextInputProps } from '@guaribas-ui/react'
+import type { Meta, StoryObj } from '@storybook/react'
+import { Box, Text, TextInput, TextInputProps } from '@healthcare-ui/react'
 
 export default {
   title: 'Form/Text Input',
   component: TextInput,
-  args: {
-    label: 'Email address',
-  },
+  args: {},
   decorators: [
     (Story) => {
       return (
@@ -14,6 +12,7 @@ export default {
           as="label"
           css={{ display: 'flex', flexDirection: 'column', gap: '$2' }}
         >
+          <Text size="sm">Email address</Text>
           {Story()}
         </Box>
       )
@@ -32,9 +31,10 @@ export const Disabled: StoryObj<TextInputProps> = {
     disabled: true,
   },
 }
+
 export const WithPrefix: StoryObj<TextInputProps> = {
   args: {
-    prefix: 'guaribas.dev/',
+    prefix: 'cal.com/',
     placeholder: 'your-username',
   },
 }