+++ packages/react/src/components/Avatar/index.tsx
@@ -1,6 +1,6 @@
 import { User } from 'phosphor-react'
 import { ComponentProps } from 'react'
-import { AvatarContainer, AvatarFallback, AvatarImage } from './styles'
+import { AvatarContainer, AvatarImage, AvatarFallback } from './styles'
 
 export interface AvatarProps extends ComponentProps<typeof AvatarImage> {}
 
@@ -8,6 +8,7 @@ export function Avatar(props: AvatarProps) {
   return (
     <AvatarContainer>
       <AvatarImage {...props} />
+
       <AvatarFallback delayMs={600}>
         <User />
       </AvatarFallback>