+++ packages/react/src/components/Avatar/styles.ts
@@ -8,12 +8,14 @@ export const AvatarContainer = styled(Avatar.Root, {
   height: '$16',
   overflow: 'hidden',
 })
+
 export const AvatarImage = styled(Avatar.Image, {
   width: '100%',
   height: '100%',
   objectFit: 'cover',
   borderRadius: 'inherit',
 })
+
 export const AvatarFallback = styled(Avatar.Fallback, {
   width: '100%',
   height: '100%',
@@ -21,7 +23,7 @@ export const AvatarFallback = styled(Avatar.Fallback, {
   alignItems: 'center',
   justifyContent: 'center',
   backgroundColor: '$gray600',
-  color: '$gray100',
+  color: '$gray800',
 
   svg: {
     width: '$6',