+++ packages/react/src/components/Box.tsx
@@ -1,4 +1,4 @@
-import { ComponentProps, ElementType } from 'react'
+import { ElementType, ComponentProps } from 'react'
 import { styled } from '../styles'
 
 export const Box = styled('div', {