+++ packages/react/src/components/Checkbox/styles.ts
@@ -1,12 +1,13 @@
 import * as Checkbox from '@radix-ui/react-checkbox'
 import { styled, keyframes } from '../../styles'
 
-export const CheckBoxContainer = styled(Checkbox.Root, {
+export const CheckboxContainer = styled(Checkbox.Root, {
   all: 'unset',
   width: '$6',
   height: '$6',
   backgroundColor: '$gray900',
   borderRadius: '$xs',
+  lineHeight: 0,
   cursor: 'pointer',
   overflow: 'hidden',
   boxSizing: 'border-box',
@@ -18,20 +19,31 @@ export const CheckBoxContainer = styled(Checkbox.Root, {
   '&[data-state="checked"]': {
     backgroundColor: '$ignite300',
   },
+
   '&:focus, &[data-state="checked"]': {
     border: '2px solid $ignite300',
   },
 })
 
 const slideIn = keyframes({
-  from: { transform: 'translateY(-100%)' },
-  to: { transform: 'translateY(0)' },
+  from: {
+    transform: 'translateY(-100%)',
+  },
+  to: {
+    transform: 'translateY(0)',
+  },
 })
+
 const slideOut = keyframes({
-  from: { transform: 'translateY(0)' },
-  to: { transform: 'translateY(-100%)' },
+  from: {
+    transform: 'translateY(0)',
+  },
+  to: {
+    transform: 'translateY(-100%)',
+  },
 })
-export const CheckBoxIndication = styled(Checkbox.Indicator, {
+
+export const CheckboxIndicator = styled(Checkbox.Indicator, {
   color: '$white',
   width: '$4',
   height: '$4',