+++ packages/react/src/components/multiStep/styles.ts
@@ -2,8 +2,10 @@ import { styled } from '../../styles'
 import { Text } from '../Text'
 
 export const MultiStepContainer = styled('div', {})
+
 export const Label = styled(Text, {
   color: '$gray200',
+
   defaultVariants: {
     size: 'xs',
   },
@@ -11,9 +13,9 @@ export const Label = styled(Text, {
 
 export const Steps = styled('div', {
   display: 'grid',
+  gridTemplateColumns: 'repeat(var(--steps-size), 1fr)',
   gap: '$2',
   marginTop: '$1',
-  gridTemplateColumns: 'repeat(var(--steps-size), 1fr)',
 })
 
 export const Step = styled('div', {