+++ packages/react/CHANGELOG.md
@@ -1,31 +1,91 @@
-# @guaribas-ui/react
+# @healthcare-ui/react
 
-## 4.0.0
+## 2.3.0
 
-### Major Changes
+### Minor Changes
 
-- fix: types and ref components
+- new package
 
-## 3.0.0
+## 2.2.2
 
-### Major Changes
+### Patch Changes
 
-- update
+- Bump version
+
+## 2.2.1
+
+### Patch Changes
+
+- Fix ref forwarding to input
+
+## 2.2.0
+
+### Minor Changes
+
+- Add the possibility to forward ref to text input
+
+## 2.1.5
+
+### Patch Changes
+
+- Fix placeholder color on text inputs
+
+## 2.1.4
+
+### Patch Changes
+
+- Add default focus to buttons and align text input content to center
+
+## 2.1.3
+
+### Patch Changes
+
+- Set default input size to medium
+
+## 2.1.2
+
+### Patch Changes
+
+- Change avatar default size to 64px
 
 ## 2.1.1
 
 ### Patch Changes
 
-- add Modal and include new token spaces
+- Keep border style when checkbox is checked
 
 ## 2.1.0
 
 ### Minor Changes
 
-- Add option to pass referencs to TextInput
+- Add size propery to TextInput
+
+## 2.0.4
+
+### Patch Changes
+
+- Increase padding on Box
+
+## 2.0.3
+
+### Patch Changes
+
+- Fix export of stitches consts
+
+## 2.0.2
+
+### Patch Changes
+
+- Export stitches consts
+
+## 2.0.1
+
+### Patch Changes
+
+- Add files published on NPM package
 
 ## 2.0.0
 
 ### Major Changes
 
-- Publish guaribas-ui
+- Add display name to all components