+++ packages/ts-config/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @healthcare-ui/eslint-config
+# @healthcare-ui/ts-config
 
 ## 1.1.0